arize_ai-phoenix/requirements.txt
llama-index==0.10.18
llama-index-llms-groq==0.1.3
chainlit==1.0.401
groq==0.4.2
python-dotenv==1.0.1
llama-index-callbacks-arize-phoenix>0.1.3
